Move getDefaultColor to ApplicationPreferences so it can be used in core modules
